requirements.txt
fastapi>=0.111,<0.112
uvicorn[standard]>=0.30,<0.31
pydantic>=2.7,<3
pydantic-settings>=2.2,<3
SQLAlchemy>=2.0,<3
alembic>=1.13,<2
cryptography>=42,<44
PyJWT>=2.8,<3
python-jose[cryptography]>=3.3,<4
orjson>=3.10,<4
httpx>=0.27,<0.28
python-dotenv>=1.0,<2
python-multipart>=0.0.9
PyNaCl>=1.5,<2.0
